refactor useOssFragment
