+++ examples/relay-hook-example/todo/package.json
@@ -16,12 +16,12 @@
     "prop-types": "^15.7.2",
     "react": "^16.9.0",
     "react-dom": "^16.8.4",
-    "relay-runtime": "7.0.0",
+    "relay-runtime": "7.1.0",
     "relay-hooks": "2.0.0-rc.4",
     "whatwg-fetch": "3.0.0",
     "isomorphic-fetch": "2.2.1",
     "es6-promise": "4.2.8",
-    "react-relay": "7.0.0"
+    "react-relay": "7.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.3.4",
@@ -47,7 +47,7 @@
     "flow-bin": "^0.94.0",
     "flow-typed": "^2.5.1",
     "prettier": "^1.16.4",
-    "relay-compiler": "^7.0.0",
+    "relay-compiler": "^7.1.0",
     "webpack": "^4.29.6",
     "webpack-dev-server": "^3.2.1"
   },
